Adjust the limit when loading the last page (AR only).
If per_page is 30 and there are 200 records after filtered (with ransack, etc)
and 2 million records overall, with a lousy database like mysql, going to page
7 (i.e. the last page) may require traversing all 2 million records if a wrong
index is chosen for the query (e.g. the index for a ORDER BY column is picked
over the index for a WHERE clause).

In order to adjust the limit, we need to always spawn the count query first to
get the total. The count query is needed anyway in most cases (as far as our
pagination usage is concerned, in all cases).

To accomodate this change, several things have been removed:

1. The `offset_value + size` optimization to get the total for the last page.
   The optimization is not applicable anymore as we always get the total first.

2. The `per_page` override. It was relying on `limit_value`, which is no longer
   a fixed value now that we adjust the limit dynamically. Plus, it causes side
   effect by calling `limit`.

3. The `limit` override. It causes side effect by calling `offset`, and does so
   using `limit_value` which is assumed to be a fixed value.

WARN: This change applies to AR only, and breaks dozens of specs.