Adjust the limit when loading the last page (AR only).
If per_page is 30 and there are 200 records after filtered (with ransack, etc)
and 2 million records overall, with a lousy database like mysql, going to page
7 (i.e. the last page) may require traversing all 2 million records if a wrong
index is chosen for the query (e.g. the index for a ORDER BY column is picked
over the index for a WHERE clause).

In order to adjust the limit, we need to always spawn the count query first to
get the total. The count query is needed anyway in most cases (as far as our
pagination usage is concerned, in all cases).